fix(decode): scale mi-size byte limit by u8 element size
Feature ID: INFRA-DECODE-SERIAL-HOT-PATHS

Follow-up to the mi-size grid narrowing: the backing store is now Vec<u8>,
but ensure_mi_size_allocation_within_limits still multiplied the element
count by size_of::<usize>(), so the MaxDecodedFrameBytes guard over-counted
this allocation 8x and could reject a tile frontier that the narrowed store
would fit. Multiply by size_of::<u8>() to match the real element width. The
two byte-budget tests are updated to the true 576-byte size and a threshold
below it so they still exercise the guard.

+++ crates/splot-decode/src/bitstream/tile_payload/tile_frontier.rs
@@ -257,7 +257,7 @@ fn ensure_mi_size_allocation_within_limits(
         .ensure_mul(
             DecodeLimitName::MaxDecodedFrameBytes,
             allocation.entry_count() as u64,
-            size_of::<usize>() as u64,
+            size_of::<u8>() as u64,
         )?
         .actual();
     limits.ensure_allocation_len(DecodeLimitName::MaxDecodedFrameBytes, allocation_bytes)?;
